Initial ECO schema upload
ECO (External Code Objects) is a collection of tables to hold fundamental metadata for well structured languages and/or code outside of the immediate SQL Server environment, but related to server or BI operations. Currently XML shredding is supported particularly for assessing metadata and dependencies related to SSRS Reports (.rdl), and SSIS Packages (.dtsx).