Initial TMP schema load
TMP (tempory) schema is a hybrid staging/non-permanent area for regular and ad-hoc analysis data to be stored. The core processes in the CAT schema use a collection of tables in this schema to stage data from multiple databases prior to processing. LIB schema process also use some of the TMP schema tables. Tables in this schema with standard prefix and hierarchy codes ("REG_0100") are routinely truncated but never dropped. Other tables may be dropped if identified as abandoned, or routinely by some processes. Privileged users should not consider data stored in this schema safe; but short-term analysis data such as value-hashes, or document shredding content may be persisted.